Embedding all datasets all Gensim 300 models implemented

+++ word_embedding/models/services/gensim/embed_text_gensim.py
@@ -6,6 +6,6 @@ def embed_text_gensim(params):
     try:
         vector = params.model.get_vector(params.word, norm=True)
     except:
-        vector = np.zeros((params.model.vector_size,))
+        vector = np.zeros((params.model.vector_size,), dtype=np.float32)
     
     return vector
